requirement.common.txt
django
psycopg2
Pillow
djangorestframework
django-filter
django-cleanup
django-webpack-loader
